+++ apollo/server/models/Person.ts
@@ -3,96 +3,37 @@
  * @module Person
  */
 
-import { Sequelize, Model, DataTypes, Optional } from "sequelize";
-import { gql } from "apollo-server";
+import { Sequelize, Model, DataTypes } from "sequelize";
+import gql from "graphql-tag";
 import { Major } from "@/models/Major";
 import { College } from "@/models/College";
-import { DateTime } from "luxon";
-
-enum GenderEnum {
-  Male = "Male",
-  Female = "Female",
-  None = "None",
-}
-
-/** The enum for the nine collges plus the option None */
-export enum CollegeEnum {
-  CC = "CC",
-  UC = "UC",
-  NA = "NA",
-  SC = "SC",
-  MC = "MC",
-  SHHO = "SHHO",
-  CW = "CW",
-  WYS = "WYS",
-  LWS = "LWS",
-  GS = "GS",
-  None = "None",
-}
-
-export enum MemberStatusEnum {
-  Unactivated = "Unactivated", // for those who have not finished registration
-  Activated = "Activated",
-  Expired = "Expired",
-}
-
-export enum RegistrationTypeEnum {
-  New = "New",
-  Renewal = "Renewal",
-}
-
-/** All the attributes in the Person model */
-export interface PersonAttributes {
-  id: number;
-  /** The student id of the student */
-  sid: string;
-  chineseName: string | null;
-  englishName: string;
-  gender: GenderEnum | null;
-  dateOfBirth: string | null;
-  email: string | null;
-  phone: string | null;
-  college: CollegeEnum | College;
-  /** The major program's code of the student */
-  major: string | Major;
-  /** The date that the student entered the university */
-  dateOfEntry: string;
-  expectedGraduationDate: string;
-  /**
-   * The date that the student became a member of the society \
-   * Null for non-member
-   */
-  memberSince: string | null;
-  /** The date that the membership expires, null for until grad */
-  memberUntil: string | null;
-  /** The status of the person */
-  status: MemberStatusEnum;
-  /** The type of the registration */
-  registrationType: RegistrationTypeEnum | null;
-}
-
-/** All the timestamps auto generated by sequelize */
-export type ModelTimestamps = {
-  createdAt: Date;
-  updatedAt: Date;
-};
-
-/** All the attributes stored in the database */
-export type PersonModelAttributes = Omit<
+import {
+  GenderEnum,
+  CollegeEnum,
+  MemberStatusEnum,
+  RegistrationTypeEnum,
   PersonAttributes,
-  "status" | "registrationType"
-> &
-  ModelTimestamps;
+  PersonModelAttributes,
+  PersonCreationAttributes,
+  statusOf,
+  registrationTypeOf,
+} from "@/utils/Person";
 
-/** All the attributes needed to create an instance of the Person model */
-export type PersonCreationAttributes = Optional<
-  Omit<PersonModelAttributes, "createdAt" | "updatedAt">,
-  "id"
->;
-
-/** All the attributes needed to update an instance of the Person model */
-export type PersonUpdateAttributes = Partial<PersonModelAttributes> &
-  Pick<PersonModelAttributes, "sid">;
+export {
+  GenderEnum,
+  CollegeEnum,
+  MemberStatusEnum,
+  RegistrationTypeEnum,
+  statusOf,
+  registrationTypeOf,
+} from "@/utils/Person";
+export type {
+  PersonAttributes,
+  ModelTimestamps,
+  PersonModelAttributes,
+  PersonCreationAttributes,
+  PersonUpdateAttributes,
+} from "@/utils/Person";
 
 /** A class for the Person model */
 export class Person
@@ -136,46 +77,12 @@ export class Person
 
   /** The status of the person */
   public get status(): MemberStatusEnum {
-    return Person.status(this);
+    return statusOf(this);
   }
 
   /** The type of the registration */
   public get registrationType(): RegistrationTypeEnum | null {
-    return Person.registrationType(this);
-  }
-
-  /** The status of the person */
-  public static status(person: PersonModelAttributes): MemberStatusEnum {
-    if (!person.memberSince) {
-      return MemberStatusEnum.Unactivated;
-    }
-    if (DateTime.fromISO(person.expectedGraduationDate) < DateTime.local()) {
-      return MemberStatusEnum.Expired;
-    }
-    if (
-      person.memberUntil &&
-      DateTime.fromISO(person.memberUntil) < DateTime.local()
-    ) {
-      return MemberStatusEnum.Expired;
-    }
-    return MemberStatusEnum.Activated;
-  }
-
-  /** The type of the registration */
-  public static registrationType(
-    person: PersonModelAttributes
-  ): RegistrationTypeEnum | null {
-    if (!person.memberSince) {
-      return RegistrationTypeEnum.New;
-    }
-    if (
-      person.memberUntil &&
-      DateTime.fromJSDate(person.updatedAt) >
-        DateTime.fromISO(person.memberUntil)
-    ) {
-      return RegistrationTypeEnum.Renewal;
-    }
-    return null;
+    return registrationTypeOf(this);
   }
 
   // timestamps!
